refactor(Modal): Split into multiple files
All subcomponents were contained in a single file, which was pretty
large. Splitting into multiple files makes it better.
