+++ react/Modal/AnimatedContentHeader.jsx
@@ -1,0 +1,9 @@
+import { Component } from 'react'
+
+class AnimatedContentHeader extends Component {
+  render() {
+    return this.props.children
+  }
+}
+
+export default AnimatedContentHeader
