+++ react/Modal/index.jsx
@@ -2,265 +2,23 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import cx from 'classnames'
 import Overlay from '../Overlay'
-import { Button } from '../Button'
-import Icon from '../Icon'
 import migrateProps from '../helpers/migrateProps'
-import palette from '../palette'
 import Portal from '../Portal'
 import uniqueId from 'lodash/uniqueId'
 import styles from './styles.styl'
-
-class AnimatedContentHeader extends Component {
-  render() {
-    return this.props.children
-  }
-}
-
-function _getChildrenToRender(children) {
-  return React.Children.map(children, child =>
-    child && child.nodeName !== AnimatedContentHeader ? child : null
-  )
-}
-
-function _getAnimatedHeader(children) {
-  return React.Children.toArray(children).find(
-    child => child && child.nodeName === AnimatedContentHeader
-  )
-}
-
-class ModalContent extends Component {
-  constructor(props) {
-    super(props)
-    const { children } = this.props
-    this.refreshComputedParts(children)
-    this.state = {
-      displayGhostHeader: false
-    }
-  }
-
-  // for preact
-  // eslint-disable-next-line react/no-deprecated
-  componentWillUpdate(nextProps, nextState) {
-    this.UNSAFE_componentWillUpdate(nextProps, nextState)
-  }
-
-  UNSAFE_componentWillUpdate = nextProps => {
-    const { children } = nextProps
-    this.refreshComputedParts(children)
-  }
-
-  refreshComputedParts(children) {
-    const animatedHeader = _getAnimatedHeader(children)
-    this.animatedHeader = animatedHeader
-    this.childrenToRender = animatedHeader
-      ? _getChildrenToRender(children)
-      : children
-  }
-
-  componentDidMount() {
-    this.scrollingContent.addEventListener('scroll', this.handleScroll, {
-      passive: true
-    })
-  }
-
-  componentWillUnmount() {
-    this.scrollingContent.removeEventListener('scroll', this.handleScroll)
-  }
-
-  handleScroll = () => {
-    if (!this.contentHeader) return
-    const { headerHeight } = this
-    if (!headerHeight && this.contentHeader.clientHeight) {
-      this.headerHeight = this.contentHeader.clientHeight
-      return
-    }
-    if (
-      !this.state.displayGhostHeader &&
-      this.scrollingContent.scrollTop > headerHeight - 20
-    ) {
-      this.setState(() => ({ displayGhostHeader: true }))
-    } else if (
-      this.state.displayGhostHeader &&
-      this.scrollingContent.scrollTop < headerHeight - 20
-    ) {
-      this.setState(() => ({ displayGhostHeader: false }))
-    }
-  }
-
-  render() {
-    const { className, fixed } = this.props
-
-    const { displayGhostHeader } = this.state
-    const { animatedHeader, childrenToRender } = this
-    return (
-      <div
-        className={cx(
-          styles[`c-modal-content${fixed ? '-fixed' : ''}`],
-          className
-        )}
-        ref={div => {
-          this.scrollingContent = div
-        }}
-      >
-        {animatedHeader && (
-          <div
-            className={animatedHeader.attributes.className}
-            ref={div => {
-              this.contentHeader = div
-            }}
-          >
-            <h2 className={styles['c-modal-illu-header']}>
-              {animatedHeader.children}
-            </h2>
-            <div
-              className={cx(
-                styles['c-modal-illu-header--ghost'],
-                animatedHeader.attributes.activeClassName,
-                {
-                  [styles['is-active']]: displayGhostHeader
-                }
-              )}
-            >
-              {animatedHeader.children}
-            </div>
-          </div>
-        )}
-        {childrenToRender}
-      </div>
-    )
-  }
-}
+import ModalContent from './ModalContent'
+import ModalSection from './ModalSection'
+import { ModalHeader, ModalBrandedHeader } from './ModalHeader'
+import ModalCross from './ModalCross'
+import ModalFooter from './ModalFooter'
+import ModalButtons from './ModalButtons'
+import AnimatedContentHeader from './AnimatedContentHeader'
 
 const ModalDescription = ModalContent
-
-const ModalSection = ({ children, className }) => (
-  <div
-    className={cx(
-      styles['c-modal-content'],
-      styles['c-modal-section'],
-      className
-    )}
-  >
-    {children}
-  </div>
-)
-
-const ModalBrandedHeader = ({ logo, bg, className, style = {} }) => (
-  <h2
-    className={cx(styles['c-modal-header--branded'], className)}
-    style={{ background: bg, ...style }}
-  >
-    <img src={logo} alt="" />
-  </h2>
-)
-
-const ModalHeader = ({
-  title,
-  children,
-  className,
-  appIcon,
-  appName,
-  appEditor,
-  style,
-  id
-}) => {
-  const isTitle = typeof children === 'string'
-  return (
-    <div
-      className={cx(styles['c-modal-header'], className)}
-      style={style}
-      id={id}
-    >
-      {title && <h2>{title}</h2>}
-      {isTitle ? <h2>{children}</h2> : children}
-      {appName && (
-        <h2 className={styles['c-modal-app']}>
-          {appIcon && (
-            <img className={styles['c-modal-app-icon']} src={appIcon} />
-          )}
-          {appEditor && (
-            <span className={styles['c-app-editor']}>
-              {appEditor}
-              &nbsp;
-            </span>
-          )}
-          {appName}
-        </h2>
-      )}
-    </div>
-  )
-}
 
 const ModalTitle = props => {
   console.log('ModalTitle is a deprecated component, use ModalHeader instead')
   return <ModalHeader {...props} />
-}
-
-const ModalCross = ({ onClick, color, className }) => (
-  <Button
-    theme="close"
-    className={cx(styles['c-modal-close'], className)}
-    onClick={onClick}
-    extension="narrow"
-    type="button"
-    label="Close"
-    iconOnly
-  >
-    <Icon
-      icon="cross"
-      width="24"
-      height="24"
-      color={color || palette['coolGrey']}
-    />
-  </Button>
-)
-
-const ModalFooter = ({ children, className }) => (
-  <div className={cx(styles['c-modal-footer'], className)}>{children}</div>
-)
-
-const ModalButtons = ({
-  secondaryText,
-  secondaryAction,
-  secondaryType,
-  primaryText,
-  primaryAction,
-  primaryType,
-  children,
-  className
-}) => (
-  <div
-    className={cx(
-      'u-flex',
-      'u-flex-wrap-reverse',
-      'u-flex-justify-end',
-      styles['c-modal-footer--button'],
-      className
-    )}
-  >
-    {children}
-    {secondaryText && secondaryAction && (
-      <Button
-        theme={secondaryType}
-        onClick={secondaryAction}
-        label={secondaryText}
-        className="u-flex-auto-s"
-      />
-    )}
-    {primaryText && primaryAction && (
-      <Button
-        theme={primaryType}
-        onClick={primaryAction}
-        label={primaryText}
-        className="u-flex-auto-s"
-      />
-    )}
-  </div>
-)
-
-ModalButtons.defaultProps = {
-  primaryType: 'regular',
-  secondaryType: 'secondary'
 }
 
 class Modal extends Component {
